pyproject.toml
[project]
name = "sonolus-demo-engine"
version = "0.1.0"
description = "Demo engine for Sonolus.py"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sonolus-py>=0.1.1",
]
